Refactor current agent class out of system
The agent didn't really do anything except hold send a request to the rest gateway. This moves the test for the gateway into the gateway and has the runner do agent's one thing for it

I didn't delete agent, though, as I want it to be the library's holder for the agent configuration. This is just a small refactor to prepare for that