Make the agent operable on its own
As demonstrated in the test, one can create an agent and send it a function to run, and it will run on it's own. This code is a duplicate of `call_function` in the AgentRunner.

Not yet tested, but an agent will invoke callbacks depending on the action type it's been given in the state map. The thinking is that this can be how the runner invokes other agents.