Have agent infer parameters from procs
When defining a tool function with a method or block, this enables the agent to infer parameters from the method or block handed in.

If the parameters are required and their names is built out from the method or proc. They will default to being strings with no enum or description

If a method is sent it to create_function with no name, the method's name will be used